Add CallCompositeCustomButtonOptions
