+++ azure-communication-ui/calling/src/main/java/com/azure/android/communication/ui/calling/models/CallCompositeButtonOptions.java
@@ -6,18 +6,11 @@ package com.azure.android.communication.ui.calling.models;
 import com.azure.android.communication.ui.calling.CallCompositeEventHandler;
 
 public class CallCompositeButtonOptions {
-    OnValueChangedListener<Integer> onDrawableIdChangedListener;
-    OnValueChangedListener<String> onTitleChangedListener;
-    OnValueChangedListener<Boolean> onVisibilityChangedListener;
-    OnValueChangedListener<Boolean> onEnabledChangedListener;
-
     private int drawableId;
     private String title;
     private CallCompositeEventHandler<CallCompositeButtonClickEvent> onClickHandler;
     private Boolean isVisible = true;
     private Boolean isEnabled = true;
-
-    private CallCompositeCustomButtonPlacement placement = CallCompositeCustomButtonPlacement.OVERFLOW;
 
     interface OnValueChangedListener<T> {
         void onValueChanged(T newValue);
@@ -29,9 +22,6 @@ public class CallCompositeButtonOptions {
 
     public CallCompositeButtonOptions setDrawableId(final int drawableId) {
         this.drawableId = drawableId;
-        if (onDrawableIdChangedListener != null) {
-            onDrawableIdChangedListener.onValueChanged(drawableId);
-        }
         return this;
     }
 
@@ -41,9 +31,6 @@ public class CallCompositeButtonOptions {
 
     public CallCompositeButtonOptions setTitle(final String title) {
         this.title = title;
-        if (onTitleChangedListener != null) {
-            onTitleChangedListener.onValueChanged(title);
-        }
         return this;
     }
 
@@ -57,24 +44,12 @@ public class CallCompositeButtonOptions {
         return this;
     }
 
-    public CallCompositeButtonOptions setPlacement(final CallCompositeCustomButtonPlacement placement) {
-        this.placement = placement;
-        return this;
-    }
-
-    public CallCompositeCustomButtonPlacement getPlacement() {
-        return this.placement;
-    }
-
     public boolean isVisible() {
         return this.isVisible;
     }
 
     public CallCompositeButtonOptions setVisibility(final boolean isVisible) {
         this.isVisible = isVisible;
-        if (onVisibilityChangedListener != null) {
-            onVisibilityChangedListener.onValueChanged(isVisible);
-        }
         return this;
     }
 
@@ -84,9 +59,6 @@ public class CallCompositeButtonOptions {
 
     public CallCompositeButtonOptions setEnable(final boolean isEnabled) {
         this.isEnabled = isEnabled;
-        if (onEnabledChangedListener != null) {
-            onEnabledChangedListener.onValueChanged(isVisible);
-        }
         return this;
     }
 }
